+++ include/md5.h
@@ -37,7 +37,8 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+#include <stdint.h>
+typedef uint32_t UINT4;
 
 void MDString (char *string, unsigned char *digest);
 void MDFileHash(char *string, char *filehash, unsigned int bytes);
